requirements.txt
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dateutil==2.9.0
rich==13.7.1
anthropic==0.49.0
python-dotenv==1.0.1
tzlocal==5.2
